Fix cookie library ignoring call order

+++ garrysmod/lua/includes/modules/cookie.lua
@@ -7,13 +7,22 @@ end
 
 module( "cookie", package.seeall )
 
-local CachedEntries = {}
-local BufferedWrites = {}
-local BufferedDeletes = {}
+local CachedEntries = {
+	--[[
+		key:
+			1 = number expireTime - if the SysTime exired then the cookie is fetched from SQL
+			2 = string cookieValue
+	]]
+}
+local BufferedQueue = {
+	--[[
+		key = value (false for delete else for insert it's a string value)
+	]]
+}
 
 local function GetCache( key )
 
-	if ( BufferedDeletes[ key ] ) then return nil end
+	if ( BufferedQueue[ key ] == false ) then return nil end
 
 	local entry = CachedEntries[ key ]
 
@@ -35,8 +44,7 @@ end
 local function FlushCache()
 
 	CachedEntries = {}
-	BufferedWrites = {}
-	BufferedDeletes = {}
+	BufferedQueue = {}
 
 end
 
@@ -44,16 +52,15 @@ local function CommitToSQLite()
 
 	sql.Begin()
 
-	for k, v in pairs( BufferedWrites ) do
-		sql.Query( "INSERT OR REPLACE INTO cookies ( key, value ) VALUES ( " .. SQLStr( k ) .. ", " .. SQLStr( v ) .. " )" )
+	for k, v in pairs( BufferedQueue ) do
+		if ( v == false ) then
+			sql.Query( "DELETE FROM cookies WHERE key = " .. SQLStr( k ) )
+		else
+			sql.Query( "INSERT OR REPLACE INTO cookies ( key, value ) VALUES ( " .. SQLStr( k ) .. ", " .. SQLStr( v ) .. " )" )
+		end
 	end
 
-	for k, v in pairs( BufferedDeletes ) do
-		sql.Query( "DELETE FROM cookies WHERE key = " .. SQLStr( k ) )
-	end
-
-	BufferedWrites = {}
-	BufferedDeletes = {}
+	BufferedQueue = {}
 
 	sql.Commit()
 
@@ -69,14 +76,21 @@ local function SetCache( key, value )
 
 	if ( value == nil ) then return Delete( key ) end
 
-	if CachedEntries[ key ] then
-		CachedEntries[ key ][ 1 ] = SysTime() + 30
-		CachedEntries[ key ][ 2 ] = value
-	else
-		CachedEntries[ key ] = { SysTime() + 30, value }
+	local strValue = tostring( value )
+
+	-- It is unlikely that this could ever happen but with a invalid __tostring method tostring may silently fail
+	if ( strValue == nil ) then
+		error( "bad argument #2 to 'value' (string expected, got " .. type( value ) .. ")", 2 )
 	end
 
-	BufferedWrites[ key ] = value
+	if CachedEntries[ key ] then
+		CachedEntries[ key ][ 1 ] = SysTime() + 30
+		CachedEntries[ key ][ 2 ] = strValue
+	else
+		CachedEntries[ key ] = { SysTime() + 30, strValue }
+	end
+
+	BufferedQueue[ key ] = strValue
 
 	ScheduleCommit()
 
@@ -106,8 +120,7 @@ end
 function Delete( name )
 
 	CachedEntries[ name ] = nil
-	BufferedWrites[ name ] = nil
-	BufferedDeletes[ name ] = true
+	BufferedQueue[ name ] = false
 
 	ScheduleCommit()
